Consolidate Union and Interface Type Resolvers into a single Abstract Type Resolver
